Exercise 15, password validation complete.
Added a Graphviz image of how the function should flow.

Added a utility function for verifying something is a string.

Built the code for the actual exercise. It doesn't obfuscate the
password as it is enterred. It seems exceptionally difficult to do
without bringing in a 3rd-party package, so I'm skipping it for now.